Refactor. js util 생성 - queryselector util -> $ 함수 구현 - fetch by movie Id 함수 구현
