+++ src/index.html
@@ -71,7 +71,7 @@
     </button>
 
     <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-element-bundle.min.js"></script>
-    <script src="/src/js/index.js" type="module"></script>
+    <script src="/src/js/main.js" type="module"></script>
   </body>
   
 </body>
