+++ src/js/helper.js
@@ -1,0 +1,3 @@
+export const $ = (selector) => {
+  return document.querySelector(selector);
+}
